requirements.txt
gitingest
openai==1.68.2
python-dotenv